Implement Phase 1: Create reply/ package with type aliases
- Create reply/ package with all 6 error types moved from transport/errors.go
- Convert transport/errors.go to type aliases pointing to reply/
- Move test from transport/errors_test.go to reply/errors_test.go
