+++ reply/errors_test.go
@@ -1,16 +1,18 @@
-package transport
+package reply_test
 
 import (
 	"errors"
 	"fmt"
+	"testing"
+
 	"github.com/duh-rpc/duh-go"
+	"github.com/kapetan-io/querator/reply"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func TestNewErrInvalid(t *testing.T) {
-	in := NewInvalidOption("invalid key")
+	in := reply.NewInvalidOption("invalid key")
 	assert.Equal(t, "invalid key", in.Error())
 	err := fmt.Errorf("wrap: %w", in)
 	assert.Equal(t, "wrap: invalid key", err.Error())
